use dao hibernate in service

+++ src/main/java/jm/task/core/jdbc/service/UserServiceImpl.java
@@ -1,14 +1,14 @@
 package jm.task.core.jdbc.service;
 
 import jm.task.core.jdbc.dao.UserDao;
-import jm.task.core.jdbc.dao.UserDaoJDBCImpl;
+import jm.task.core.jdbc.dao.UserDaoHibernateImpl;
 import jm.task.core.jdbc.model.User;
 
 import java.util.List;
 
 public class UserServiceImpl implements UserService {
 
-    private UserDao userDao = new UserDaoJDBCImpl();
+    private UserDao userDao = new UserDaoHibernateImpl();
 
     public void createUsersTable() {
         userDao.createUsersTable();
@@ -19,10 +19,16 @@ public class UserServiceImpl implements UserService {
     }
 
     public void saveUser(String name, String lastName, byte age) {
+        if (name == null || name.length() == 0 || lastName == null || lastName.length() == 0 || age < 0) {
+            return;
+        }
         userDao.saveUser(name, lastName, age);
     }
 
     public void removeUserById(long id) {
+        if (id < 0) {
+            return;
+        }
         userDao.removeUserById(id);
     }
